support entitlement check and patch

+++ arm-rhel-was-nd-cluster/src/main/scripts/install.sh
@@ -22,8 +22,8 @@ create_dmgr_profile() {
     adminUserName=$5
     adminPassword=$6
 
-    /opt/IBM/WebSphere/ND/V9/bin/manageprofiles.sh -create -profileName ${profileName} -hostName $hostName \
-        -templatePath /opt/IBM/WebSphere/ND/V9/profileTemplates/management -serverType DEPLOYMENT_MANAGER \
+    ${tWASDirectory}/bin/manageprofiles.sh -create -profileName ${profileName} -hostName $hostName \
+        -templatePath ${tWASDirectory}/profileTemplates/management -serverType DEPLOYMENT_MANAGER \
         -nodeName ${nodeName} -cellName ${cellName} -enableAdminSecurity true -adminUserName ${adminUserName} -adminPassword ${adminPassword}
 }
 
@@ -31,7 +31,7 @@ add_admin_credentials_to_soap_client_props() {
     profileName=$1
     adminUserName=$2
     adminPassword=$3
-    soapClientProps=/opt/IBM/WebSphere/ND/V9/profiles/${profileName}/properties/soap.client.props
+    soapClientProps=${tWASDirectory}/profiles/${profileName}/properties/soap.client.props
 
     # Add admin credentials
     sed -i "s/com.ibm.SOAP.securityEnabled=false/com.ibm.SOAP.securityEnabled=true/g" "$soapClientProps"
@@ -39,7 +39,7 @@ add_admin_credentials_to_soap_client_props() {
     sed -i "s/com.ibm.SOAP.loginPassword=/com.ibm.SOAP.loginPassword=${adminPassword}/g" "$soapClientProps"
 
     # Encrypt com.ibm.SOAP.loginPassword
-    /opt/IBM/WebSphere/ND/V9/profiles/${profileName}/bin/PropFilePasswordEncoder.sh "$soapClientProps" com.ibm.SOAP.loginPassword
+    ${tWASDirectory}/profiles/${profileName}/bin/PropFilePasswordEncoder.sh "$soapClientProps" com.ibm.SOAP.loginPassword
 }
 
 create_systemd_service() {
@@ -47,19 +47,21 @@ create_systemd_service() {
     srvDescription=$2
     profileName=$3
     serverName=$4
-    srvPath=/etc/systemd/system/${srvName}.service
 
     # Add systemd unit file
-    echo "[Unit]" > "$srvPath"
-    echo "Description=${srvDescription}" >> "$srvPath"
-    echo "[Service]" >> "$srvPath"
-    echo "Type=forking" >> "$srvPath"
-    echo "ExecStart=/opt/IBM/WebSphere/ND/V9/profiles/${profileName}/bin/startServer.sh ${serverName}" >> "$srvPath"
-    echo "ExecStop=/opt/IBM/WebSphere/ND/V9/profiles/${profileName}/bin/stopServer.sh ${serverName}" >> "$srvPath"
-    echo "PIDFile=/opt/IBM/WebSphere/ND/V9/profiles/${profileName}/logs/${serverName}/${serverName}.pid" >> "$srvPath"
-    echo "SuccessExitStatus=143 0" >> "$srvPath"
-    echo "[Install]" >> "$srvPath"
-    echo "WantedBy=default.target" >> "$srvPath"
+    cat <<EOF > /etc/systemd/system/${srvName}.service
+[Unit]
+Description=${srvDescription}
+RequiresMountsFor=/datadrive
+[Service]
+Type=forking
+ExecStart=/bin/sh -c "${tWASDirectory}/profiles/${profileName}/bin/startServer.sh ${serverName}"
+ExecStop=/bin/sh -c "${tWASDirectory}/profiles/${profileName}/bin/stopServer.sh ${serverName}"
+PIDFile=${tWASDirectory}/profiles/${profileName}/logs/${serverName}/${serverName}.pid
+SuccessExitStatus=143 0
+[Install]
+WantedBy=default.target
+EOF
 
     # Enable service
     systemctl daemon-reload
@@ -74,13 +76,13 @@ create_cluster() {
     members=$5
     dynamic=$6
 
-    nodes=( $(/opt/IBM/WebSphere/ND/V9/profiles/${profileName}/bin/wsadmin.sh -lang jython -c "AdminConfig.list('Node')" \
+    nodes=( $(${tWASDirectory}/profiles/${profileName}/bin/wsadmin.sh -lang jython -c "AdminConfig.list('Node')" \
         | grep -Po "(?<=\/nodes\/)[^|]*(?=|.*)" | grep -v $dmgrNode | sed 's/^/"/;s/$/"/') )
     while [ ${#nodes[@]} -ne $members ]
     do
         sleep 5
         echo "adding more nodes..."
-        nodes=( $(/opt/IBM/WebSphere/ND/V9/profiles/${profileName}/bin/wsadmin.sh -lang jython -c "AdminConfig.list('Node')" \
+        nodes=( $(${tWASDirectory}/profiles/${profileName}/bin/wsadmin.sh -lang jython -c "AdminConfig.list('Node')" \
             | grep -Po "(?<=\/nodes\/)[^|]*(?=|.*)" | grep -v $dmgrNode | sed 's/^/"/;s/$/"/') )
     done
     sleep 60
@@ -91,7 +93,7 @@ create_cluster() {
         sed -i "s/\${CLUSTER_NAME}/${clusterName}/g" create-dcluster.py
         sed -i "s/\${NODE_GROUP_NAME}/DefaultNodeGroup/g" create-dcluster.py
         sed -i "s/\${CORE_GROUP_NAME}/DefaultCoreGroup/g" create-dcluster.py
-        /opt/IBM/WebSphere/ND/V9/profiles/${profileName}/bin/wsadmin.sh -lang jython -f create-dcluster.py
+        ${tWASDirectory}/profiles/${profileName}/bin/wsadmin.sh -lang jython -f create-dcluster.py
     else
         echo "all nodes are managed, creating cluster..."
         nodes_string=$( IFS=,; echo "${nodes[*]}" )
@@ -99,7 +101,7 @@ create_cluster() {
         sed -i "s/\${CELL_NAME}/${cellName}/g" create-cluster.py
         sed -i "s/\${CLUSTER_NAME}/${clusterName}/g" create-cluster.py
         sed -i "s/\${NODES_STRING}/${nodes_string}/g" create-cluster.py
-        /opt/IBM/WebSphere/ND/V9/profiles/${profileName}/bin/wsadmin.sh -lang jython -f create-cluster.py
+        ${tWASDirectory}/profiles/${profileName}/bin/wsadmin.sh -lang jython -f create-cluster.py
     fi
 
     echo "cluster \"${clusterName}\" is successfully created!"
@@ -124,16 +126,16 @@ create_custom_profile() {
     sleep 60
     echo "dmgr is ready to add nodes"
 
-    output=$(/opt/IBM/WebSphere/ND/V9/bin/manageprofiles.sh -create -profileName $profileName -hostName $hostName -nodeName $nodeName \
-        -profilePath /opt/IBM/WebSphere/ND/V9/profiles/$profileName -templatePath /opt/IBM/WebSphere/ND/V9/profileTemplates/managed \
+    output=$(${tWASDirectory}/bin/manageprofiles.sh -create -profileName $profileName -hostName $hostName -nodeName $nodeName \
+        -profilePath ${tWASDirectory}/profiles/$profileName -templatePath ${tWASDirectory}/profileTemplates/managed \
         -dmgrHost $dmgrHostName -dmgrPort $dmgrPort -dmgrAdminUserName $dmgrAdminUserName -dmgrAdminPassword $dmgrAdminPassword 2>&1)
     while echo $output | grep -qv "SUCCESS"
     do
         sleep 10
         echo "adding node failed, retry it later..."
-        rm -rf /opt/IBM/WebSphere/ND/V9/profiles/$profileName
-        output=$(/opt/IBM/WebSphere/ND/V9/bin/manageprofiles.sh -create -profileName $profileName -hostName $hostName -nodeName $nodeName \
-            -profilePath /opt/IBM/WebSphere/ND/V9/profiles/$profileName -templatePath /opt/IBM/WebSphere/ND/V9/profileTemplates/managed \
+        rm -rf ${tWASDirectory}/profiles/$profileName
+        output=$(${tWASDirectory}/bin/manageprofiles.sh -create -profileName $profileName -hostName $hostName -nodeName $nodeName \
+            -profilePath ${tWASDirectory}/profiles/$profileName -templatePath ${tWASDirectory}/profileTemplates/managed \
             -dmgrHost $dmgrHostName -dmgrPort $dmgrPort -dmgrAdminUserName $dmgrAdminUserName -dmgrAdminPassword $dmgrAdminPassword 2>&1)
     done
     echo $output
@@ -162,16 +164,47 @@ while getopts "m:c:f:h:r:x:" opt; do
     esac
 done
 
+# Check whether the user is entitled or not
+while [ ! -f "/var/log/cloud-init-was.log" ]
+do
+    echo "waiting for was entitlement check started..."
+    sleep 5
+done
+
+isDone=false
+while [ $isDone = false ]
+do
+    result=`(tail -n1) </var/log/cloud-init-was.log`
+    if [[ $result = Unentitled ]] || [[ $result = Entitled ]]; then
+        isDone=true
+    else
+        echo "waiting for was entitlement check completed..."
+        sleep 5
+    fi
+done
+echo $result
+
+# Terminate the process for the un-entitled user
+if [ ${result} = Unentitled ]; then
+    exit 1
+fi
+
+# Update applications installed on the system
+yum update
+
 # Turn off firewall
 systemctl stop firewalld
 systemctl disable firewalld
+
+# WebSphere installation directory pre-set by the base image
+tWASDirectory=/datadrive/IBM/WebSphere/ND/V9
 
 # Create cluster by creating deployment manager, node agent & add nodes to be managed
 if [ "$dmgr" = True ]; then
     create_dmgr_profile Dmgr001 $(hostname) Dmgr001Node Dmgr001NodeCell "$adminUserName" "$adminPassword"
     add_admin_credentials_to_soap_client_props Dmgr001 "$adminUserName" "$adminPassword"
     create_systemd_service was_dmgr "IBM WebSphere Application Server ND Deployment Manager" Dmgr001 dmgr
-    /opt/IBM/WebSphere/ND/V9/profiles/Dmgr001/bin/startServer.sh dmgr
+    ${tWASDirectory}/profiles/Dmgr001/bin/startServer.sh dmgr
     create_cluster Dmgr001 Dmgr001Node Dmgr001NodeCell MyCluster $members $dynamic
 else
     create_custom_profile Custom $(hostname) $(hostname)Node01 $dmgrHostName 8879 "$adminUserName" "$adminPassword"
